platform: syntacore: reserve opensbi memory
OpenSBI is loaded to the first 0x200000 bytes of DRAM. However OpenSBI
is an implementation of SBI firmware. So, unlike U-Boot bootloader, it
must stay alive in DRAM throughout the lifetime of the system in order
to serve SBI requests from Linux kernel. So reserve the first 0x200000
bytes of memory, carving them out from the Linux memory map.

Signed-off-by: Sergey Matyukevich <sergey.matyukevich@syntacore.com>